# server/api/handlers.go
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"log"
	"net/http"
	"strings"
	"strconv"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/cors"
	gotdtg "github.com/gotd/td/tg"
	httpSwagger "github.com/swaggo/http-swagger"
	"omnigram/db"
	"omnigram/tg"
	"omnigram/auth"

	// Docs generated by swag
	_ "omnigram/docs"
)

type Server struct {
	Repo      *db.Repository
	TgManager *tg.Manager
}

func NewServer(repo *db.Repository, manager *tg.Manager) *Server {
	return &Server{
		Repo:      repo,
		TgManager: manager,
	}
}

func (s *Server) Router() http.Handler {
	r := chi.NewRouter()
	r.Use(middleware.Logger)
	r.Use(cors.Handler(cors.Options{
		AllowedOrigins:   []string{"http://localhost:5173", "http://127.0.0.1:5173"},
		AllowedMethods:   []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
		AllowedHeaders:   []string{"Accept", "Authorization", "Content-Type", "X-CSRF-Token", "X-Session-ID"},
		ExposedHeaders:   []string{"Link"},
		AllowCredentials: true,
		MaxAge:           300,
	}))
	r.Use(auth.SessionIDMiddleware) // Global: all API requests must have X-Session-ID

	r.Get("/swagger/*", httpSwagger.WrapHandler)

	r.Route("/api", func(r chi.Router) {
		r.Route("/auth", func(r chi.Router) {
			r.Post("/send-code", s.SendCode)
			r.Post("/verify-code", s.VerifyCode)
			r.Post("/verify-password", s.VerifyPassword)
			r.Group(func(r chi.Router) {
				r.Use(auth.Middleware)
				r.Get("/me", s.GetMe)
				r.Post("/logout", s.Logout)
			})
		})

		r.Group(func(r chi.Router) {
			r.Use(auth.Middleware)
			r.Get("/channels", s.GetChannels)
			r.Get("/channels/{id}/messages", s.GetMessages)
			r.Post("/grab", s.GrabChannel)
			r.Get("/dialogs", s.GetDialogs)
			r.Post("/grab_by_id", s.GrabByID)
			r.Post("/grab_invite", s.GrabByInvite)
			r.Post("/search", s.SearchMessages)
			r.Post("/enrich", s.EnrichDialog)
			r.Get("/enrich/tasks", s.GetEnrichmentTasks)
		})
	})

	return r
}

type SearchRequest struct {
	Query string `json:"query"`
}

type SearchResponse struct {
	Analysis string       `json:"analysis"`
	Results  []db.Message `json:"results"`
}

type GrabRequest struct {
	Username string `json:"username"`
}

type GrabByIDRequest struct {
	ChatID int64 `json:"chat_id"`
}

type GrabByInviteRequest struct {
	Hash string `json:"hash"`
}

type EnrichRequest struct {
	DialogID int64 `json:"dialog_id"`
}

// GetChannels godoc
// @Summary List all grabbed channels
// @Description Get a list of all channels that have been grabbed
// @Tags channels
// @Produce json
// @Success 200 {array} db.Channel
// @Router /api/channels [get]
func (s *Server) GetChannels(w http.ResponseWriter, r *http.Request) {
	channels, err := s.Repo.GetChannels()
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(channels)
}

// GetMessages godoc
// @Summary List messages for a channel
// @Description Get a list of messages for a specific channel ID
// @Tags messages
// @Produce json
// @Param id path int true "Channel ID"
// @Success 200 {array} db.Message
// @Router /api/channels/{id}/messages [get]
func (s *Server) GetMessages(w http.ResponseWriter, r *http.Request) {
	idStr := chi.URLParam(r, "id")
	id, err := strconv.ParseInt(idStr, 10, 64)
	if err != nil {
		http.Error(w, "Invalid channel ID", http.StatusBadRequest)
		return
	}

	messages, err := s.Repo.GetMessages(id)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(messages)
}

// GrabChannel godoc
// @Summary Grab a channel
// @Description Starts downloading channel history in the background
// @Tags channels
// @Accept json
// @Produce json
// @Param request body api.GrabRequest true "Channel Username"
// @Success 202 {string} string "Started grabbing in background"
// @Router /api/grab [post]
func (s *Server) GrabChannel(w http.ResponseWriter, r *http.Request) {
	var req GrabRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	if req.Username == "" {
		http.Error(w, "Username is required", http.StatusBadRequest)
		return
	}

	sessionID := auth.GetSessionID(r.Context())
	app, err := s.TgManager.GetApp(r.Context(), sessionID)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	// Run in background...
	go func() {
		err := app.GrabChannel(context.Background(), req.Username)
		if err != nil {
			log.Printf("Background grab failed for session %s: %v", sessionID, err)
		}
	}()

	w.WriteHeader(http.StatusAccepted)
	w.Write([]byte("Started grabbing in background"))
}

// GetDialogs godoc
// @Summary List recent dialogs
// @Description Get a list of up to 100 recent dialogs to find Chat IDs
// @Tags dialogs
// @Produce json
// @Success 200 {array} object
// @Router /api/dialogs [get]
func (s *Server) GetDialogs(w http.ResponseWriter, r *http.Request) {
	app, err := s.TgManager.GetApp(r.Context(), auth.GetSessionID(r.Context()))
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	type DialogResponse struct {
		ID    int64  `json:"id"`
		Title string `json:"title"`
		Type  string `json:"type"`
	}

	chats, users, err := app.GetDialogs(r.Context())
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	var resp []DialogResponse
	for _, chatClass := range chats {
		switch c := chatClass.(type) {
		case *gotdtg.Chat:
			resp = append(resp, DialogResponse{ID: c.ID, Title: c.Title, Type: "group"})
		case *gotdtg.Channel:
			resp = append(resp, DialogResponse{ID: c.ID, Title: c.Title, Type: "supergroup/channel"})
		}
	}

	for _, userClass := range users {
		if u, ok := userClass.(*gotdtg.User); ok {
			// Don't skip deleted users, but label them
			if u.Deleted {
				resp = append(resp, DialogResponse{ID: u.ID, Title: "Deleted User", Type: "user"})
				continue
			}

			name := strings.TrimSpace(u.FirstName + " " + u.LastName)
			if name == "" {
				if u.Username != "" {
					name = "@" + u.Username
				} else {
					name = "User " + fmt.Sprint(u.ID)
				}
			}
			
			userType := "user"
			if u.Bot {
				userType = "bot"
			}
			
			resp = append(resp, DialogResponse{ID: u.ID, Title: name, Type: userType})
		}
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(resp)
}

// GrabByID godoc
// @Summary Grab a group by ID
// @Description Starts downloading channel/group history in the background by ID
// @Tags channels
// @Accept json
// @Produce json
// @Param request body api.GrabByIDRequest true "Chat ID"
// @Success 202 {string} string "Started grabbing by ID in background"
// @Router /api/grab_by_id [post]
func (s *Server) GrabByID(w http.ResponseWriter, r *http.Request) {
	var req GrabByIDRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	sessionID := auth.GetSessionID(r.Context())
	app, err := s.TgManager.GetApp(r.Context(), sessionID)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	go func() {
		err := app.GrabByID(context.Background(), req.ChatID)
		if err != nil {
			log.Printf("Background grab by ID failed for session %s: %v", sessionID, err)
		}
	}()

	w.WriteHeader(http.StatusAccepted)
	w.Write([]byte("Started grabbing by ID in background"))
}

// GrabByInvite godoc
// @Summary Grab a group by invite hash
// @Description Joins and starts downloading channel/group history in the background by invite link hash
// @Tags channels
// @Accept json
// @Produce json
// @Param request body api.GrabByInviteRequest true "Invite Hash (e.g. AbCdEfGhIjK from t.me/+AbCdEfGhIjK)"
// @Success 202 {string} string "Started grabbing by Invite in background"
// @Router /api/grab_invite [post]
func (s *Server) GrabByInvite(w http.ResponseWriter, r *http.Request) {
	var req GrabByInviteRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	sessionID := auth.GetSessionID(r.Context())
	app, err := s.TgManager.GetApp(r.Context(), sessionID)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	go func() {
		err := app.GrabByInvite(context.Background(), req.Hash)
		if err != nil {
			log.Printf("Background grab by invite failed for session %s: %v", sessionID, err)
		}
	}()

	w.WriteHeader(http.StatusAccepted)
	w.Write([]byte("Started grabbing by Invite in background"))
}

// SearchMessages godoc
// @Summary Search and analyze messages
// @Description Performs vector search for relevant messages and uses Gemini to analyze them
// @Tags search
// @Accept json
// @Produce json
// @Param request body api.SearchRequest true "Search Query"
// @Success 200 {object} api.SearchResponse
// @Router /api/search [post]
func (s *Server) SearchMessages(w http.ResponseWriter, r *http.Request) {
	var req SearchRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	if req.Query == "" {
		http.Error(w, "Query is required", http.StatusBadRequest)
		return
	}

	app, err := s.TgManager.GetApp(r.Context(), auth.GetSessionID(r.Context()))
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	if app.Embedder == nil {
		http.Error(w, "Embedder is not initialized (OLLAMA_URL missing)", http.StatusInternalServerError)
		return
	}

	// 1. Embed the query
	queryEmbedding, err := app.Embedder.EmbedText(r.Context(), req.Query)
	if err != nil {
		http.Error(w, fmt.Sprintf("Failed to embed query: %v", err), http.StatusInternalServerError)
		return
	}

	// 2. Search for relevant messages
	results, err := s.Repo.SearchMessages(queryEmbedding, 100) // Top 100 results
	if err != nil {
		http.Error(w, fmt.Sprintf("Search failed: %v", err), http.StatusInternalServerError)
		return
	}

	if len(results) == 0 {
		w.Header().Set("Content-Type", "application/json")
		json.NewEncoder(w).Encode(SearchResponse{
			Analysis: "No relevant messages found.",
			Results:  []db.Message{},
		})
		return
	}

	// 3. Decrypt results for analysis
	var messageTexts []string
	for i := range results {
		// Prefer context if available, otherwise just original message
		var toDecrypt []byte = results[i].EncryptedData
		var nonce []byte = results[i].Nonce
		
		if results[i].EncryptedContext != nil {
			toDecrypt = results[i].EncryptedContext
			nonce = results[i].NonceContext
		}

		if toDecrypt != nil {
			decrypted, err := auth.Decrypt(app.EncryptionKey, toDecrypt, nonce)
			if err != nil {
				log.Printf("Failed to decrypt message/context %d: %v", results[i].ID, err)
				results[i].ContextText = "[Decryption Failed]"
			} else {
				results[i].ContextText = string(decrypted)
				results[i].Text = string(decrypted) // Fallback for Text field
			}
		}
		messageTexts = append(messageTexts, results[i].ContextText)
	}

	// 4. Analyze with Gemini
	analysis, err := app.Gemini.AnalyzeMessages(r.Context(), req.Query, messageTexts)
	if err != nil {
		log.Printf("Gemini analysis failed: %v", err)
		analysis = "Search completed, but analysis failed."
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(SearchResponse{
		Analysis: analysis,
		Results:  results,
	})
}

// EnrichDialog godoc
// @Summary Enrich a dialog
// @Description Starts enrichment process (encryption + windowed embeddings) in background
// @Tags channels
// @Accept json
// @Produce json
// @Param request body api.EnrichRequest true "Dialog ID"
// @Success 202 {string} string "Started enrichment in background"
// @Router /api/enrich [post]
func (s *Server) EnrichDialog(w http.ResponseWriter, r *http.Request) {
	var req EnrichRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	sessionID := auth.GetSessionID(r.Context())
	
	// Just mark the task as processing in the DB. 
	// The background worker in Manager will pick it up and process batches.
	err := s.Repo.UpsertEnrichmentTask(req.DialogID, sessionID, "processing", 0)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusAccepted)
	w.Write([]byte("Task marked for enrichment. Background worker will process it in batches."))
}

// GetEnrichmentTasks godoc
// @Summary List enrichment tasks
// @Description Get a list of all enrichment tasks and their progress for current session
// @Tags channels
// @Produce json
// @Success 200 {array} db.EnrichmentTask
// @Router /api/enrich/tasks [get]
func (s *Server) GetEnrichmentTasks(w http.ResponseWriter, r *http.Request) {
	sessionID := auth.GetSessionID(r.Context())
	tasks, err := s.Repo.GetAllEnrichmentTasksBySession(sessionID)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(tasks)
}